tests/wpt/tests/html/canvas/offscreen/compositing/2d.composite.globalAlpha.invalid.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.composite.globalAlpha.invalid</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.composite.globalAlpha.invalid</h1>
<p class="desc"></p>


<script>
var t = async_test("");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.globalAlpha = 0.5;
  // This may not set it to exactly 0.5 if it is rounded/quantised, so
  // remember for future comparisons.
  var a = ctx.globalAlpha;
  ctx.globalAlpha = Infinity;
  _assertSame(ctx.globalAlpha, a, "ctx.globalAlpha", "a");
  ctx.globalAlpha = -Infinity;
  _assertSame(ctx.globalAlpha, a, "ctx.globalAlpha", "a");
  ctx.globalAlpha = NaN;
  _assertSame(ctx.globalAlpha, a, "ctx.globalAlpha", "a");
  t.done();

});
</script>
